requirements.txt
absl-py==2.2.2
aiofiles==24.1.0
anyio==4.9.0
argcomplete==3.6.2
attrs==25.3.0
beautifulsoup4==4.13.3
boto3==1.37.31
botocore==1.37.31
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dacite==1.7.0
Deprecated==1.2.18
dill==0.3.9
dnspython==2.7.0
fiftyone==1.4.1
fiftyone-brain==0.20.1
fiftyone_db==1.1.7
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
ftfy==6.3.1
future==1.0.0
glob2==0.7
graphql-core==3.2.6
grpcio==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.12.2
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
inflate64==1.0.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
kaleido==0.2.1
kiwisolver==1.4.8
lazy_loader==0.4
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
mongoengine==0.29.1
motor==3.6.1
mpmath==1.3.0
multivolumefile==0.2.3
narwhals==1.34.1
networkx==3.4.2
numpy==2.2.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pip==25.0.1
plotly==6.0.1
pprintpp==0.4.0
priority==2.0.0
protobuf==6.30.2
psutil==7.0.0
py7zr==0.22.0
pybcj==1.0.3
pycocotools==2.0.8
pycryptodomex==3.22.0
pydash==8.0.5
pymongo==4.9.2
pyparsing==3.2.3
pyppmd==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzstd==0.16.2
rarfile==4.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rtree==1.4.0
s3transfer==0.11.4
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sse-starlette==0.10.3
sseclient-py==1.8.0
starlette==0.46.1
strawberry-graphql==0.263.2
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.3.30
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
universal-analytics-python3==1.1.1
urllib3==2.4.0
voxel51-eta==0.14.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xmltodict==0.14.2